Redesigning settings page, working on organizer events
- redesigning settings page, still in the works
- organizer events does not populate user-specific events, still in the works
