+++ Volunteer/Volunteer.xcworkspace/xcuserdata/bri.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -46,8 +46,8 @@
             filePath = "Events/AllEventsTableViewController.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "188"
-            endingLineNumber = "188">
+            startingLineNumber = "180"
+            endingLineNumber = "180">
          </BreakpointContent>
       </BreakpointProxy>
    </Breakpoints>
